+++ src/routes/irbs/index.js
@@ -4,9 +4,14 @@
 import PrincetonCovidDisinformationIRB from "./PrincetonCovidDisinformationIRB.svelte";
 import RS01Consent from "./RS01Consent.svelte";
 import StanfordBeyondThePaywallIRB from "./StanfordBeyondThePaywallIRB.svelte";
+import FacebookPixelHuntConsent from "./FacebookPixelHuntConsent.svelte";
+import PrincetonSearchUsageIRB from "./PrincetonSearchUsageIRB.svelte";
 
 export default {
     "rally.news.study@princeton.edu": PrincetonCovidDisinformationIRB,
     "rally-study-01@mozilla.org": RS01Consent,
-    "beyond-the-paywall@rally.mozilla.org": StanfordBeyondThePaywallIRB
+    "beyond-the-paywall@rally.mozilla.org": StanfordBeyondThePaywallIRB,
+    "facebook-pixel-hunt@rally.mozilla.org": FacebookPixelHuntConsent,
+    "search-engine-usage@rally.mozilla.org": PrincetonSearchUsageIRB,
+    "princeton-political-and-covid-19-news-study@rally.mozilla.org": PrincetonCovidDisinformationIRB,
 }